refactor: PointV1Controller.java 구현으로 포인트 URI 수정
